frontend/requirements.txt
requests==2.20.0
Pillow==5.4.1
numpy==1.15.0
